Merged revision(s) 518 from branches/SFML2:
Bugfixes:
- Slightly improved error handling (not fully merged from branches/SFML2)
- Fixed Credits GUI broken in r492
- Updated some copyright dates to 2014

Refactorizations:
- Fixed some warnings
- Avoid copying TScore objects
- Small simplifications
........

Fixed mistake in previous merge commit.

+++ src/config_screen.cpp
@@ -171,7 +171,7 @@ void CGameConfig::Enter() {
 
 	int siz = FT.AutoSizeN(5);
 	textbuttons[0] = AddTextButton(Trans.Text(28), area.left+50, AutoYPosN(80), siz);
-	double len = FT.GetTextWidth(Trans.Text(8));
+	float len = FT.GetTextWidth(Trans.Text(8));
 	textbuttons[1] = AddTextButton(Trans.Text(15), area.right-len-50, AutoYPosN(80), siz);
 
 	Music.Play(param.config_music, -1);
