+++ src/event.cpp
@@ -160,7 +160,7 @@ void CEvent::Enter() {
 	ResetGUI();
 	int siz = FT.AutoSizeN(5);
 	textbuttons[1] = AddTextButton(Trans.Text(8), area.left + 100, AutoYPosN(80), siz);
-	double len = FT.GetTextWidth(Trans.Text(13));
+	float len = FT.GetTextWidth(Trans.Text(13));
 	textbuttons[0] = AddTextButton(Trans.Text(13), area.right -len - 100, AutoYPosN(80), siz);
 	textbuttons[2] = AddTextButton(Trans.Text(15), CENTER, AutoYPosN(80), siz);
